+++ google2/src/test/scala/org/broadinstitute/dsde/workbench/google2/Generators.scala
@@ -44,9 +44,12 @@ object Generators {
     topic <- genTopicName
     filters <- genFilters
   } yield NotificationRequest(topic, "JSON_API_V1", filters.eventTypes, filters.objectNamePrefix)
+  val genDiskName = alphaLowerStrOfLength(10).map(DiskName)
+
+  def alphaLowerStrOfLength(n: Int): Gen[String] = Gen.listOfN(n, Gen.alphaLowerChar).map(_.mkString)
 
   implicit val arbProjectTopicName: Arbitrary[TopicName] = Arbitrary(genTopicName)
   implicit val arbNotificationResponse: Arbitrary[NotificationResponse] = Arbitrary(genNotificationResponse)
   implicit val arbNotificationRequest: Arbitrary[NotificationRequest] = Arbitrary(genNotificationRequest)
-
+  implicit val arbDiskName: Arbitrary[DiskName] = Arbitrary(genDiskName)
 }